+++ webui/package.json
@@ -32,6 +32,7 @@
     "react-redux": "^5.0.4",
     "react-redux-firebase": "^1.5.0-alpha",
     "react-redux-multilingual": "^1.0.4",
+    "react-responsive": "^1.3.4",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
     "react-scripts": "0.9.5",
